Migrated to new adapter
